Added logs + checks

+++ install.sh
@@ -11,34 +11,65 @@ olddir=~/.dotfiles_old                 			# old dotfiles backup directory
 files="vimrc gitconfig bash_profile gitignore"  # list of files/folders to symlink in homedir
 vimdir=~/.vim/bundle
 github=https://github.com
+tempdotfiles="$TMPDIR/dotfiles"
 
 function install_vim_package {
-  package=$1
-  folder=$2
-  if [ ! -d "$vimdir/$folder" ]; then
-    git clone "$github/$package" "$vimdir/$folder"
+  link=$1
+  package=$2
+  dir=`pwd`
+  if [ ! -d "$vimdir/$package" ]; then
+    echo "Cloning $package"
+    git clone "$github/$link" "$vimdir/$package" >> "$tempdotfiles/clone_$package"
+  else
+    cd "$vimdir/$package"
+    echo "Pulling updates for $package"
+    git pull >> "$tempdotfiles/pull_$package"
+    cd $dir
   fi
 }
 
-##########
+for i in "$@"
+do
+case $i in
+	-b=|--backup)
+    backup=1
+    shift
+    ;;
+    *)
+    ;;
+esac
+done
 
-# create dotfiles_old in homedir
-echo "Creating $olddir for backup of any existing dotfiles in ~"
-mkdir -p $olddir
-echo "...done"
+if [ $backup ] && [ ! -d "$olddir" ]; then
+	echo "Creating $olddir for backup of any existing dotfiles in ~"
+	mkdir -p $olddir
+	echo "...done"
+fi
 
-# change to the dotfiles directory
 echo "Changing to the $dir directory"
+mkdir -p $tempdotfiles
 cd $dir
-# move any existing dotfiles in homedir to dotfiles_old directory, then create symlinks 
 for file in $files; do
-    echo "Moving any existing dotfiles from ~ to $olddir"
-    mv ~/.$file $olddir
-    echo "Creating symlink to $file in home directory."
-    ln -s $dir/$file ~/.$file
+    if [ $backup ]; then
+    	echo "Moving any existing dotfiles from ~ to $olddir"
+    	mv ~/.$file $olddir
+    fi
+    if [ -e ~/.$file ]; then
+      echo "~/.$file exists"
+      if [ ! $backup ] && [ ! -L ~/.$file ]; then
+        echo "Backup options in not selected and it's not a symlink, so bailing out. re-run with -b option"
+        exit 1
+      else
+        echo "File backed up or symlink, so removing it"
+        rm ~/.$file
+      fi
+	    echo "Creating symlink to $file in home directory."
+    	ln -s $dir/$file ~/.$file
+	 fi
 done
 
 # reinstall vim plugins
+mkdir -p $vimdir
 install_vim_package bling/vim-airline vim-airline
 install_vim_package mhinz/vim-signify vim-signify
 install_vim_package rking/ag.vim ag
@@ -52,3 +83,5 @@ install_vim_package Shougo/neocomplete.vim.git neocomplete.vim
 install_vim_package ervandew/supertab.git supertab
 install_vim_package SirVer/ultisnips.git ultisnips
 install_vim_package majutsushi/tagbar.git tagbar
+
+echo "Logs available at $tempdotfiles"
